install go in /opt/go
it makes more sense for the installation of go to keep most
files in the same directory, because this is intended by the
authors.  by installing go to /opt/go, we enable a lot of
functionality: godoc, goinstall, cgo and all that stuff
works correctly out of the box.

the reason for not installing go to /usr/lib/go is simple.
go has a lot of files that do usually not belong to lib, but
since they must be kept together /usr/lib/go is not
appropriate and i think /opt/go is a good target for an aur
package. it is not necessary that this is also true for a
[community] package.

a good thingt is that the go installation supports a
separate bin directory so we can install the binaries in
/usr/bin and everything works as expected.

the current installation now is roughly 150 megabytes,
that's huge, but in my opinion it is affordable and this way
we can keep the PKGBUILD quite simple.

some editor related files (currently emacs and vim) are
still installed in the corresponding global directories:
- /usr/share/emacs/site-lisp/go-mode{,-load}.el
- /usr/share/vim/vimfiles/{ftdetect,syntax}/go.vim

the license file is installed in /usr/share/licenses/go.

bash completion and shell profile are still installed in
/etc/bash_completion.d/go and /etc/profile.d/go.sh
respectively.

that's it with the big changes, have fun!